Add files via upload
THIS IS THE WORKING PROJECT. cmake and functions 1-4 work and pass their tests. There are some tests that are in their domain that don't pass due to unfinished functions being used in those tests. For example, the following allocate_and_free test does not pass because the block_store_release function is not built yet. Once that is built, I believe the tests will pass because the functions that are built passed their other tests. Let me know if that's not the case. Good luck fellas

TEST(block_store_alloc_free_req, allocate_and_free) {
    block_store_t *bs = NULL;
    bs = block_store_create();
    ASSERT_NE(nullptr, bs) << "block_store_create returned NULL when it should not have\n";
    size_t id = 100;
    id = block_store_allocate(bs);
    ASSERT_EQ(0, id) << "The id returned should be zero; this should be the first allocation.\n";
    block_store_release(bs, id);
    id = block_store_allocate(bs);
    ASSERT_EQ(0, id) << "The id returned should again be zero; did your release not work?\n";
    block_store_destroy(bs);

    score += 5;
}